web-application/requirements-web.txt
tensorflow==2.18.0
Flask==3.1.0
pillow==11.1.0